fix a massive performance degradation.
if a parser is optional, don't register a failure if it doesn't match! this was causing high branching because it would treat a failed optional as both a success and a failure. the failure would cascade backward to force it to try other paths. for example, "there's no whitespace after the + so maybe this isn't an addition at all!" ughhhhhhhhh

(also, allow a null-default optional parser to be cached. this is often used to drop optional whitespace.)