this new test is a good example of why it won't work. (more below)
the idea was: try to store the failed ParseTask in a failed match, so that if we report a failure, we have the task that failed, and you can try to "recover" by pretending that it succeeded and resuming the parse.

but i don't think this covers enough cases. many "random chars" bugs will cause the parse to end with "expected end", which is not something you can really recover from -- you need an earlier parser to skip over it. also some parsers (like optional or repeat) only report the success cases, not the failures... so if the "best" recovery is to succeed an optional that was omitted, or to extend a repeat by one more than it actually matched... there's no way to do that.

i'm shelving this and leaving it alone.